Untested changes to make Orlando easier on us
For a Vision shot to happen, the driver needs to hold the shoot button
the entire time. If they release prematurely, the Vision shot will be
aborted.

Using the new enabledInit() function in Robot.java

Using the new enabledInit() and disabledInit() functions in the
Subsystems

Cleanup and organization of code

Moved control of the Vision tracking light from IntakeCommand to
ShooterCommand

DriveSubsystem.gyroTurn() now returns true or false based off of if it
reached it's timeout.