Preparing for St. Louis
VisionTargeting now constantly checks if the Robot is alligned with the
goal. If it is not, it provides data as to which direction the Robot
should move in order to get there. This data is displayed on
SmartDashboard and available programmedly.

Automated "vision shooting" is now done by having the Robot turn in the
direction of the goal until it is properly alligned.

Adjusted high goal auto to be ready for testing and more safe/accurate
and put TODOs on the other incomplete autonomouses.

Shooter revving now defaults to the high goal rate.

Re-enabled gyro turn code and added a warning about using it when going
over bumps.

There are now seperate functions for having encoder-based driving
correct it's angle after completing - encoderDriveGyro() and
inchDriveGyro()

Commented out unused subsystems and chooser options.

Refactored "bad ball" and "good ball" variables into "high goal" and
"low goal".

If the setting of a Camera property is unsuccessful, a warning message
is logged.

Deleted old code.

Maybe something else I've forgotten.